remove chess match functionality
cutechess can do the same, but better
