+++ bin/ai.rs
@@ -1,0 +1,18 @@
+use futures_util::{future::BoxFuture, stream::BoxStream};
+use lib::chess::{Move, Position};
+use lib::search::{Limits, Pv};
+
+/// Trait for types that know how to analyze chess [`Position`]s.
+pub trait Ai {
+    /// Finds the best [`Move`].
+    fn play<'a, 'b, 'c>(&'a mut self, pos: &'b Position, limits: Limits) -> BoxFuture<'c, Move>
+    where
+        'a: 'c,
+        'b: 'c;
+
+    /// Analyzes a [`Position`].
+    fn analyze<'a, 'b, 'c>(&'a mut self, pos: &'b Position, limits: Limits) -> BoxStream<'c, Pv>
+    where
+        'a: 'c,
+        'b: 'c;
+}
