+++ bin/applet.rs
@@ -3,13 +3,11 @@ use clap::Subcommand;
 use derive_more::From;
 
 mod analyze;
-mod play;
 mod uci;
 
 #[derive(From, Subcommand)]
 pub enum Applet {
     Analyze(analyze::Analyze),
-    Play(play::Play),
     Uci(uci::Uci),
 }
 
@@ -23,7 +21,6 @@ impl Applet {
     pub async fn execute(self) -> Result<(), Anyhow> {
         match self {
             Applet::Analyze(a) => Ok(a.execute().await?),
-            Applet::Play(a) => Ok(a.execute().await?),
             Applet::Uci(a) => Ok(a.execute().await?),
         }
     }
