--- bin/engine/uci.rs
@@ -1,679 +1,0 @@
-use crate::{io::Io, play::Play};
-use anyhow::{Context, Error as Anyhow};
-use async_trait::async_trait;
-use derive_more::{DebugCustom, Display, Error, From};
-use lib::chess::{Color, Move, Position};
-use lib::search::Limits;
-use std::{collections::HashMap, fmt::Debug, future::Future, io, pin::Pin};
-use tokio::{runtime, task::block_in_place};
-use tracing::{debug, error, instrument};
-use vampirc_uci::{self as uci, UciFen, UciMessage, UciSearchControl};
-
-pub type UciOptions = HashMap<String, Option<String>>;
-
-#[derive(DebugCustom)]
-#[debug(bound = "T: Debug")]
-#[debug(fmt = "Lazy({})")]
-enum Lazy<T, E> {
-    #[debug(fmt = "{_0:?}")]
-    Initialized(T),
-    #[debug(fmt = "?")]
-    Uninitialized(Pin<Box<dyn Future<Output = Result<T, E>> + Send + 'static>>),
-}
-
-impl<T, E> Lazy<T, E> {
-    async fn get_or_init(&mut self) -> Result<&mut T, E> {
-        match self {
-            Lazy::Initialized(v) => Ok(v),
-            Lazy::Uninitialized(f) => {
-                *self = Lazy::Initialized(f.await?);
-                match self {
-                    Lazy::Initialized(v) => Ok(v),
-                    Lazy::Uninitialized(_) => unreachable!(),
-                }
-            }
-        }
-    }
-}
-
-/// The reason why an [`Move`] could not be received from the UCI server.
-#[derive(Debug, Display, Error, From)]
-#[display(fmt = "the UCI server encountered an error")]
-pub struct UciError(#[from(forward)] io::Error);
-
-/// A Universal Chess Interface client for a chess engine.
-#[derive(Debug)]
-pub struct Uci<T: Io> {
-    io: Lazy<T, UciError>,
-}
-
-impl<T: Io + Send + 'static> Uci<T> {
-    /// Constructs [`Uci`] with the given [`UciOptions`].
-    pub fn new(mut io: T, options: UciOptions) -> Self {
-        Uci {
-            io: Lazy::Uninitialized(Box::pin(async move {
-                io.send(&UciMessage::Uci.to_string()).await?;
-                io.flush().await?;
-
-                while !matches!(uci::parse_one(io.recv().await?.trim()), UciMessage::UciOk) {}
-
-                for (name, value) in options {
-                    let set_option = UciMessage::SetOption { name, value };
-                    io.send(&set_option.to_string()).await?;
-                }
-
-                io.send(&UciMessage::UciNewGame.to_string()).await?;
-                io.send(&UciMessage::IsReady.to_string()).await?;
-                io.flush().await?;
-
-                while !matches!(uci::parse_one(io.recv().await?.trim()), UciMessage::ReadyOk) {}
-
-                Ok(io)
-            })),
-        }
-    }
-
-    async fn go(&mut self, pos: &Position, limits: Limits) -> Result<(), UciError> {
-        let fen = uci_position(pos);
-        let ctrl = uci_search_control(limits, pos.turn());
-
-        let io = self.io.get_or_init().await?;
-        io.send(&fen.to_string()).await?;
-        io.send(&ctrl.to_string()).await?;
-        io.flush().await?;
-
-        Ok(())
-    }
-}
-
-fn uci_position(pos: &Position) -> UciMessage {
-    UciMessage::Position {
-        startpos: false,
-        fen: Some(UciFen(pos.to_string())),
-        moves: Vec::new(),
-    }
-}
-
-fn uci_search_control(limits: Limits, turn: Color) -> UciMessage {
-    match limits {
-        Limits::None => UciMessage::go(),
-        Limits::Depth(d) => UciMessage::Go {
-            search_control: Some(UciSearchControl::depth(d.get())),
-            time_control: None,
-        },
-        Limits::Time(t) => UciMessage::go_movetime(
-            uci::Duration::from_std(t).unwrap_or_else(|_| uci::Duration::max_value()),
-        ),
-        Limits::Clock(t, i) => {
-            let t = uci::Duration::from_std(t).unwrap_or_else(|_| uci::Duration::max_value());
-            let i = uci::Duration::from_std(i).unwrap_or_else(|_| uci::Duration::max_value());
-
-            let [(white_time, white_increment), (black_time, black_increment)] = match turn {
-                Color::White => [(Some(t), Some(i)), (None, None)],
-                Color::Black => [(None, None), (Some(t), Some(i))],
-            };
-
-            UciMessage::Go {
-                search_control: None,
-                time_control: Some(uci::UciTimeControl::TimeLeft {
-                    white_time,
-                    black_time,
-                    white_increment,
-                    black_increment,
-                    moves_to_go: None,
-                }),
-            }
-        }
-    }
-}
-
-impl<T: Io> Drop for Uci<T> {
-    #[instrument(level = "trace", skip(self))]
-    fn drop(&mut self) {
-        let result: Result<(), Anyhow> = block_in_place(|| {
-            runtime::Handle::try_current()?.block_on(async {
-                let io = self.io.get_or_init().await?;
-                io.send(&UciMessage::Stop.to_string()).await?;
-                io.send(&UciMessage::Quit.to_string()).await?;
-                io.flush().await?;
-                Ok(())
-            })
-        });
-
-        if let Err(e) = result.context("failed to gracefully shutdown the uci engine") {
-            error!("{:?}", e);
-        }
-    }
-}
-
-#[async_trait]
-impl<T: Io + Send + 'static> Play for Uci<T> {
-    type Error = UciError;
-
-    #[instrument(level = "debug", skip(self, pos), ret(Display), err, fields(%pos, %limits))]
-    async fn play(&mut self, pos: &Position, limits: Limits) -> Result<Move, Self::Error> {
-        self.go(pos, limits).await?;
-        let io = self.io.get_or_init().await?;
-
-        loop {
-            if let UciMessage::BestMove { best_move: m, .. } =
-                uci::parse_one(io.recv().await?.trim())
-            {
-                break Ok(m.into());
-            }
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::io::MockIo;
-    use mockall::Sequence;
-    use proptest::prelude::*;
-    use std::future::ready;
-    use test_strategy::proptest;
-    use tokio::runtime;
-
-    fn any_uci_message() -> impl Strategy<Value = UciMessage> {
-        prop_oneof![
-            Just(UciMessage::Uci),
-            Just(UciMessage::UciOk),
-            Just(UciMessage::UciNewGame),
-            Just(UciMessage::IsReady),
-            Just(UciMessage::ReadyOk),
-            Just(UciMessage::Stop),
-            Just(UciMessage::Quit),
-            Just(UciMessage::PonderHit),
-            any::<(Move, Option<Move>)>().prop_map(|(m, p)| UciMessage::BestMove {
-                best_move: m.into(),
-                ponder: p.map(Into::into),
-            }),
-            any::<(Option<String>, Option<String>)>()
-                .prop_map(|(name, author)| UciMessage::Id { name, author }),
-            any::<(bool, Option<String>, Option<String>)>()
-                .prop_map(|(later, name, code)| UciMessage::Register { later, name, code }),
-            any::<(String, Option<String>)>()
-                .prop_map(|(name, value)| UciMessage::SetOption { name, value }),
-            any::<bool>().prop_map(UciMessage::Debug),
-        ]
-    }
-
-    #[proptest]
-    fn new_schedules_engine_for_lazy_initialization(o: UciOptions) {
-        assert!(matches!(
-            Uci::new(MockIo::new(), o),
-            Uci {
-                io: Lazy::Uninitialized(_),
-            }
-        ));
-    }
-
-    #[proptest]
-    fn engine_is_lazily_initialized_with_the_options_configured(
-        l: Limits,
-        o: UciOptions,
-        pos: Position,
-        m: Move,
-    ) {
-        let rt = runtime::Builder::new_multi_thread().build()?;
-        let mut io = MockIo::new();
-        let mut seq = Sequence::new();
-
-        io.expect_send()
-            .once()
-            .in_sequence(&mut seq)
-            .withf(|msg| msg == UciMessage::Uci.to_string())
-            .returning(|_| Box::pin(ready(Ok(()))));
-
-        io.expect_flush()
-            .once()
-            .in_sequence(&mut seq)
-            .returning(|| Box::pin(ready(Ok(()))));
-
-        io.expect_recv()
-            .once()
-            .in_sequence(&mut seq)
-            .returning(move || Box::pin(ready(Ok(UciMessage::UciOk.to_string()))));
-
-        for (name, value) in o.clone() {
-            let set_option = UciMessage::SetOption { name, value };
-            io.expect_send()
-                .once()
-                .in_sequence(&mut seq)
-                .withf(move |msg| msg == set_option.to_string())
-                .returning(|_| Box::pin(ready(Ok(()))));
-        }
-
-        io.expect_send()
-            .once()
-            .in_sequence(&mut seq)
-            .withf(|msg| msg == UciMessage::UciNewGame.to_string())
-            .returning(|_| Box::pin(ready(Ok(()))));
-
-        io.expect_send()
-            .once()
-            .in_sequence(&mut seq)
-            .withf(|msg| msg == UciMessage::IsReady.to_string())
-            .returning(|_| Box::pin(ready(Ok(()))));
-
-        io.expect_flush()
-            .once()
-            .in_sequence(&mut seq)
-            .returning(|| Box::pin(ready(Ok(()))));
-
-        io.expect_recv()
-            .once()
-            .in_sequence(&mut seq)
-            .returning(move || Box::pin(ready(Ok(UciMessage::ReadyOk.to_string()))));
-
-        io.expect_send().returning(|_| Box::pin(ready(Ok(()))));
-        io.expect_flush().returning(|| Box::pin(ready(Ok(()))));
-        io.expect_recv()
-            .once()
-            .returning(move || Box::pin(ready(Ok(UciMessage::best_move(m.into()).to_string()))));
-
-        let mut uci = Uci::new(io, o);
-
-        assert_eq!(
-            rt.block_on(uci.play(&pos, l))
-                .map_err(|UciError(e)| e.kind()),
-            Ok(m)
-        );
-    }
-
-    #[proptest]
-    fn initialization_ignores_invalid_uci_messages(
-        l: Limits,
-        o: UciOptions,
-        pos: Position,
-        m: Move,
-        #[filter(matches!(uci::parse_one(#msg.trim()), UciMessage::Unknown(_, _)))] msg: String,
-    ) {
-        let rt = runtime::Builder::new_multi_thread().build()?;
-        let mut io = MockIo::new();
-
-        io.expect_send().returning(|_| Box::pin(ready(Ok(()))));
-        io.expect_flush().returning(|| Box::pin(ready(Ok(()))));
-
-        io.expect_recv()
-            .once()
-            .return_once(move || Box::pin(ready(Ok(msg))));
-
-        io.expect_recv()
-            .once()
-            .returning(move || Box::pin(ready(Ok(UciMessage::UciOk.to_string()))));
-
-        io.expect_recv()
-            .once()
-            .returning(move || Box::pin(ready(Ok(UciMessage::ReadyOk.to_string()))));
-
-        io.expect_recv()
-            .once()
-            .returning(move || Box::pin(ready(Ok(UciMessage::best_move(m.into()).to_string()))));
-
-        let mut uci = Uci::new(io, o);
-
-        assert_eq!(
-            rt.block_on(uci.play(&pos, l))
-                .map_err(|UciError(e)| e.kind()),
-            Ok(m)
-        );
-    }
-
-    #[proptest]
-    fn initialization_ignores_unexpected_uci_messages(
-        l: Limits,
-        o: UciOptions,
-        pos: Position,
-        m: Move,
-        #[filter(!matches!(#msg, UciMessage::UciOk))]
-        #[strategy(any_uci_message())]
-        msg: UciMessage,
-    ) {
-        let rt = runtime::Builder::new_multi_thread().build()?;
-        let mut io = MockIo::new();
-
-        io.expect_send().returning(|_| Box::pin(ready(Ok(()))));
-        io.expect_flush().returning(|| Box::pin(ready(Ok(()))));
-
-        io.expect_recv()
-            .once()
-            .return_once(move || Box::pin(ready(Ok(msg.to_string()))));
-
-        io.expect_recv()
-            .once()
-            .returning(move || Box::pin(ready(Ok(UciMessage::UciOk.to_string()))));
-
-        io.expect_recv()
-            .once()
-            .returning(move || Box::pin(ready(Ok(UciMessage::ReadyOk.to_string()))));
-
-        io.expect_recv()
-            .once()
-            .returning(move || Box::pin(ready(Ok(UciMessage::best_move(m.into()).to_string()))));
-
-        let mut uci = Uci::new(io, o);
-
-        assert_eq!(
-            rt.block_on(uci.play(&pos, l))
-                .map_err(|UciError(e)| e.kind()),
-            Ok(m)
-        );
-    }
-
-    #[proptest]
-    fn initialization_can_fail(l: Limits, o: UciOptions, pos: Position, e: io::Error) {
-        let rt = runtime::Builder::new_multi_thread().build()?;
-        let mut io = MockIo::new();
-
-        let kind = e.kind();
-        io.expect_send()
-            .once()
-            .return_once(move |_| Box::pin(ready(Err(e))));
-
-        io.expect_send().returning(|_| Box::pin(ready(Ok(()))));
-        io.expect_flush().returning(|| Box::pin(ready(Ok(()))));
-
-        let mut uci = Uci::new(io, o);
-
-        assert_eq!(
-            rt.block_on(uci.play(&pos, l))
-                .map_err(|UciError(e)| e.kind()),
-            Err(kind)
-        );
-    }
-
-    #[proptest]
-    fn drop_gracefully_quits_initialized_engine() {
-        let rt = runtime::Builder::new_multi_thread().build()?;
-        let mut io = MockIo::new();
-
-        let mut seq = Sequence::new();
-
-        io.expect_send()
-            .once()
-            .in_sequence(&mut seq)
-            .withf(|msg| msg == UciMessage::Stop.to_string())
-            .returning(|_| Box::pin(ready(Ok(()))));
-
-        io.expect_send()
-            .once()
-            .in_sequence(&mut seq)
-            .withf(|msg| msg == UciMessage::Quit.to_string())
-            .returning(|_| Box::pin(ready(Ok(()))));
-
-        io.expect_flush()
-            .once()
-            .in_sequence(&mut seq)
-            .returning(|| Box::pin(ready(Ok(()))));
-
-        rt.block_on(async move {
-            drop(Uci {
-                io: Lazy::Initialized(io),
-            });
-        })
-    }
-
-    #[proptest]
-    fn drop_gracefully_quits_uninitialized_engine() {
-        let rt = runtime::Builder::new_multi_thread().build()?;
-        let mut io = MockIo::new();
-
-        io.expect_send()
-            .times(3)
-            .returning(|_| Box::pin(ready(Ok(()))));
-        io.expect_flush()
-            .times(2)
-            .returning(|| Box::pin(ready(Ok(()))));
-
-        io.expect_recv()
-            .once()
-            .returning(move || Box::pin(ready(Ok(UciMessage::UciOk.to_string()))));
-
-        io.expect_recv()
-            .once()
-            .returning(move || Box::pin(ready(Ok(UciMessage::ReadyOk.to_string()))));
-
-        let mut seq = Sequence::new();
-
-        io.expect_send()
-            .once()
-            .in_sequence(&mut seq)
-            .withf(|msg| msg == UciMessage::Stop.to_string())
-            .returning(|_| Box::pin(ready(Ok(()))));
-
-        io.expect_send()
-            .once()
-            .in_sequence(&mut seq)
-            .withf(|msg| msg == UciMessage::Quit.to_string())
-            .returning(|_| Box::pin(ready(Ok(()))));
-
-        io.expect_flush()
-            .once()
-            .in_sequence(&mut seq)
-            .returning(|| Box::pin(ready(Ok(()))));
-
-        rt.block_on(async move {
-            drop(Uci::new(io, UciOptions::default()));
-        })
-    }
-
-    #[proptest]
-    fn drop_recovers_from_errors(e: io::Error) {
-        let rt = runtime::Builder::new_multi_thread().build()?;
-        let mut io = MockIo::new();
-        io.expect_send()
-            .once()
-            .return_once(move |_| Box::pin(ready(Err(e))));
-
-        rt.block_on(async move {
-            drop(Uci {
-                io: Lazy::Initialized(io),
-            });
-        })
-    }
-
-    #[proptest]
-    fn drop_recovers_from_missing_runtime() {
-        drop(Uci {
-            io: Lazy::Initialized(MockIo::new()),
-        });
-    }
-
-    #[proptest]
-    fn go_instructs_engine_search_for_best_move(l: Limits, pos: Position) {
-        let rt = runtime::Builder::new_multi_thread().build()?;
-        let mut io = MockIo::new();
-        let mut seq = Sequence::new();
-
-        let fen = uci_position(&pos);
-
-        io.expect_send()
-            .once()
-            .in_sequence(&mut seq)
-            .withf(move |msg| msg == fen.to_string())
-            .returning(|_| Box::pin(ready(Ok(()))));
-
-        let ctrl = uci_search_control(l, pos.turn());
-
-        io.expect_send()
-            .once()
-            .in_sequence(&mut seq)
-            .withf(move |msg| msg == ctrl.to_string())
-            .returning(|_| Box::pin(ready(Ok(()))));
-
-        io.expect_flush()
-            .once()
-            .in_sequence(&mut seq)
-            .returning(|| Box::pin(ready(Ok(()))));
-
-        let mut uci = Uci {
-            io: Lazy::Initialized(io),
-        };
-
-        assert_eq!(
-            rt.block_on(uci.go(&pos, l)).map_err(|UciError(e)| e.kind()),
-            Ok(())
-        );
-    }
-
-    #[proptest]
-    fn go_can_fail_writing(l: Limits, pos: Position, e: io::Error) {
-        let rt = runtime::Builder::new_multi_thread().build()?;
-        let mut io = MockIo::new();
-
-        let kind = e.kind();
-        io.expect_send()
-            .once()
-            .return_once(move |_| Box::pin(ready(Err(e))));
-
-        io.expect_send().returning(|_| Box::pin(ready(Ok(()))));
-        io.expect_flush().returning(|| Box::pin(ready(Ok(()))));
-
-        let mut uci = Uci {
-            io: Lazy::Initialized(io),
-        };
-
-        assert_eq!(
-            rt.block_on(uci.go(&pos, l)).map_err(|UciError(e)| e.kind()),
-            Err(kind)
-        );
-    }
-
-    #[proptest]
-    fn go_can_fail_flushing(l: Limits, pos: Position, e: io::Error) {
-        let rt = runtime::Builder::new_multi_thread().build()?;
-        let mut io = MockIo::new();
-
-        io.expect_send().returning(|_| Box::pin(ready(Ok(()))));
-
-        let kind = e.kind();
-        io.expect_flush()
-            .once()
-            .return_once(move || Box::pin(ready(Err(e))));
-        io.expect_flush().returning(|| Box::pin(ready(Ok(()))));
-
-        let mut uci = Uci {
-            io: Lazy::Initialized(io),
-        };
-
-        assert_eq!(
-            rt.block_on(uci.go(&pos, l)).map_err(|UciError(e)| e.kind()),
-            Err(kind)
-        );
-    }
-
-    #[proptest]
-    fn search_waits_for_engine_to_find_best_move(l: Limits, pos: Position, m: Move) {
-        let rt = runtime::Builder::new_multi_thread().build()?;
-        let mut io = MockIo::new();
-
-        io.expect_send().returning(|_| Box::pin(ready(Ok(()))));
-        io.expect_flush().returning(|| Box::pin(ready(Ok(()))));
-
-        io.expect_recv()
-            .once()
-            .returning(move || Box::pin(ready(Ok(UciMessage::best_move(m.into()).to_string()))));
-
-        let mut uci = Uci {
-            io: Lazy::Initialized(io),
-        };
-
-        assert_eq!(
-            rt.block_on(uci.play(&pos, l))
-                .map_err(|UciError(e)| e.kind()),
-            Ok(m)
-        );
-    }
-
-    #[proptest]
-    fn search_ignores_invalid_uci_messages(
-        l: Limits,
-        pos: Position,
-        m: Move,
-        #[filter(matches!(uci::parse_one(#msg.trim()), UciMessage::Unknown(_, _)))] msg: String,
-    ) {
-        let rt = runtime::Builder::new_multi_thread().build()?;
-        let mut io = MockIo::new();
-
-        io.expect_send().returning(|_| Box::pin(ready(Ok(()))));
-        io.expect_flush().returning(|| Box::pin(ready(Ok(()))));
-
-        io.expect_recv()
-            .once()
-            .return_once(move || Box::pin(ready(Ok(msg))));
-
-        io.expect_recv()
-            .once()
-            .returning(move || Box::pin(ready(Ok(UciMessage::best_move(m.into()).to_string()))));
-
-        let mut uci = Uci {
-            io: Lazy::Initialized(io),
-        };
-
-        assert_eq!(
-            rt.block_on(uci.play(&pos, l))
-                .map_err(|UciError(e)| e.kind()),
-            Ok(m)
-        );
-    }
-
-    #[proptest]
-    fn search_ignores_unexpected_uci_messages(
-        l: Limits,
-        pos: Position,
-        m: Move,
-        #[filter(!matches!(#msg, UciMessage::BestMove { .. }))]
-        #[strategy(any_uci_message())]
-        msg: UciMessage,
-    ) {
-        let rt = runtime::Builder::new_multi_thread().build()?;
-        let mut io = MockIo::new();
-
-        io.expect_send().returning(|_| Box::pin(ready(Ok(()))));
-        io.expect_flush().returning(|| Box::pin(ready(Ok(()))));
-
-        io.expect_recv()
-            .once()
-            .return_once(move || Box::pin(ready(Ok(msg.to_string()))));
-
-        io.expect_recv()
-            .once()
-            .returning(move || Box::pin(ready(Ok(UciMessage::best_move(m.into()).to_string()))));
-
-        let mut uci = Uci {
-            io: Lazy::Initialized(io),
-        };
-
-        assert_eq!(
-            rt.block_on(uci.play(&pos, l))
-                .map_err(|UciError(e)| e.kind()),
-            Ok(m)
-        );
-    }
-
-    #[proptest]
-    fn search_can_fail_reading(l: Limits, pos: Position, e: io::Error) {
-        let rt = runtime::Builder::new_multi_thread().build()?;
-        let mut io = MockIo::new();
-
-        io.expect_send().returning(|_| Box::pin(ready(Ok(()))));
-        io.expect_flush().returning(|| Box::pin(ready(Ok(()))));
-
-        let kind = e.kind();
-        io.expect_recv()
-            .once()
-            .return_once(move || Box::pin(ready(Err(e))));
-
-        let mut uci = Uci {
-            io: Lazy::Initialized(io),
-        };
-
-        assert_eq!(
-            rt.block_on(uci.play(&pos, l))
-                .map_err(|UciError(e)| e.kind()),
-            Err(kind)
-        );
-    }
-}
