+++ bin/main.rs
@@ -1,14 +1,11 @@
 use anyhow::Error as Anyhow;
 use clap::Parser;
 
-mod analyze;
+mod ai;
 mod applet;
-mod build;
 mod cli;
 mod engine;
-mod game;
 mod io;
-mod play;
 
 #[tokio::main]
 async fn main() -> Result<(), Anyhow> {
