+++ lib/chess.rs
@@ -3,25 +3,21 @@ mod fen;
 mod file;
 mod r#move;
 mod outcome;
-mod pgn;
 mod piece;
 mod position;
 mod promotion;
 mod rank;
 mod role;
-mod san;
 mod square;
 
 pub use color::*;
 pub use fen::*;
 pub use file::*;
 pub use outcome::*;
-pub use pgn::*;
 pub use piece::*;
 pub use position::*;
 pub use promotion::*;
 pub use r#move::*;
 pub use rank::*;
 pub use role::*;
-pub use san::*;
 pub use square::*;
