+++ lib/chess/outcome.rs
@@ -8,12 +8,6 @@ pub enum Outcome {
     #[display(fmt = "checkmate by the {_0} player")]
     Checkmate(Color),
 
-    #[display(fmt = "{_0} player lost on time")]
-    LossOnTime(Color),
-
-    #[display(fmt = "{_0} player resigned")]
-    Resignation(Color),
-
     #[display(fmt = "stalemate")]
     Stalemate,
 
@@ -34,16 +28,13 @@ impl Outcome {
 
     /// Whether the outcome is a decisive and one of the sides has won.
     pub fn is_decisive(&self) -> bool {
-        use Outcome::*;
-        matches!(self, Checkmate(_) | LossOnTime(_) | Resignation(_))
+        matches!(self, Outcome::Checkmate(_))
     }
 
     /// The winning side, if the outcome is [decisive](`Self::is_decisive`).
     pub fn winner(&self) -> Option<Color> {
         match *self {
             Outcome::Checkmate(c) => Some(c),
-            Outcome::LossOnTime(c) => Some(!c),
-            Outcome::Resignation(c) => Some(!c),
             _ => None,
         }
     }
@@ -72,15 +63,5 @@ mod tests {
     #[proptest]
     fn side_that_checkmates_wins(c: Color) {
         assert_eq!(Outcome::Checkmate(c).winner(), Some(c));
-    }
-
-    #[proptest]
-    fn side_that_runs_out_of_time_loses(c: Color) {
-        assert_eq!(Outcome::LossOnTime(c).winner(), Some(!c));
-    }
-
-    #[proptest]
-    fn side_that_resigns_loses(c: Color) {
-        assert_eq!(Outcome::Resignation(c).winner(), Some(!c));
     }
 }
